refactor(A2-2064): move action banners with dashboard functions
