+++ packages/forms-web-app/src/lib/dashboard-functions.js
@@ -10,11 +10,16 @@ const {
 } = require('@pins/common/src/lib/format-address');
 const { formatDateForDisplay } = require('@pins/common/src/lib/format-date');
 const { caseTypeNameWithDefault } = require('@pins/common/src/lib/format-case-type');
-const { APPEAL_REPRESENTATION_STATUS } = require('pins-data-model');
-const { APPEAL_USER_ROLES, REPRESENTATION_TYPES } = require('@pins/common/src/constants');
+const {
+	APPEAL_USER_ROLES,
+	REPRESENTATION_TYPES,
+	LPA_USER_ROLE
+} = require('@pins/common/src/constants');
+const { deadlineHasPassed } = require('./deadline-has-passed');
 const logger = require('#lib/logger');
 
 /**
+ * @typedef {import('@pins/common/src/constants').AppealToUserRoles | "LPAUser" | null} UserRole
  * @typedef {import('appeals-service-api').Api.AppealCaseDetailed} AppealCaseDetailed
  * @typedef {import('appeals-service-api').Api.AppealSubmission} AppealSubmission
  * @typedef {import('appeals-service-api').Api.Representation} Representation
@@ -54,7 +59,7 @@ const {
 } = require('./full-appeal/map-planning-application');
 const { mapTypeCodeToAppealId } = require('@pins/common');
 const { businessRulesDeadline } = require('./calculate-deadline');
-const { APPEAL_CASE_STATUS } = require('pins-data-model');
+const { APPEAL_CASE_STATUS, APPEAL_REPRESENTATION_STATUS } = require('pins-data-model');
 const { calculateDaysSinceInvalidated } = require('./calculate-days-since-invalidated');
 
 const questionnaireBaseUrl = '/manage-appeals/questionnaire';
@@ -363,6 +368,16 @@ const isQuestionnaireDue = (appealCaseData) => {
 };
 
 /**
+ * @param {AppealCaseDetailed} caseData
+ * @param {UserRole} userType
+ * @returns {boolean}
+ */
+const shouldDisplayQuestionnaireDueNotification = (caseData, userType) =>
+	userType === LPA_USER_ROLE &&
+	!caseData.lpaQuestionnaireSubmittedDate &&
+	!!caseData.lpaQuestionnaireDueDate;
+
+/**
  * @param {AppealCaseDetailed} appealCaseData return object from database call
  * @returns {boolean}
  */
@@ -371,6 +386,21 @@ const isLPAStatementDue = (appealCaseData) => {
 		!!appealCaseData.statementDueDate &&
 		!appealCaseData.LPAStatementSubmittedDate &&
 		appealCaseData.caseStatus === APPEAL_CASE_STATUS.STATEMENTS
+	);
+};
+
+/**
+ * @param {AppealCaseDetailed} caseData
+ * @param {UserRole} userType
+ * @returns {boolean}
+ */
+const shouldDisplayStatementsDueBannerLPA = (caseData, userType) => {
+	return (
+		!!caseData.statementDueDate &&
+		userType === LPA_USER_ROLE &&
+		deadlineHasPassed(caseData.lpaQuestionnaireDueDate) &&
+		!caseData.LPAStatementSubmittedDate &&
+		!deadlineHasPassed(caseData.statementDueDate)
 	);
 };
 
@@ -387,6 +417,22 @@ const isLPAFinalCommentDue = (appealCaseData) => {
 };
 
 /**
+ * @param {AppealCaseDetailed} caseData
+ * @param {UserRole} userType
+ * @returns {boolean}
+ */
+const shouldDisplayFinalCommentsDueBannerLPA = (caseData, userType) => {
+	return (
+		!!caseData.finalCommentsDueDate &&
+		userType === LPA_USER_ROLE &&
+		deadlineHasPassed(caseData.statementDueDate) &&
+		deadlineHasPassed(caseData.interestedPartyRepsDueDate) &&
+		!deadlineHasPassed(caseData.finalCommentsDueDate) &&
+		!caseData.LPACommentsSubmittedDate
+	);
+};
+
+/**
  * @param {AppealCaseDetailed} appealCaseData return object from database call
  * @returns {boolean}
  */
@@ -395,6 +441,23 @@ const isLPAProofsOfEvidenceDue = (appealCaseData) => {
 		!!appealCaseData.proofsOfEvidenceDueDate &&
 		!appealCaseData.LPAProofsSubmittedDate &&
 		appealCaseData.caseStatus === APPEAL_CASE_STATUS.EVIDENCE
+	);
+};
+
+/**
+ * @param {AppealCaseDetailed} caseData
+ * @param {UserRole} userType
+ * @returns {boolean}
+ */
+const shouldDisplayProofEvidenceDueBannerLPA = (caseData, userType) => {
+	return (
+		!!caseData.proofsOfEvidenceDueDate &&
+		userType === LPA_USER_ROLE &&
+		deadlineHasPassed(caseData.statementDueDate) &&
+		deadlineHasPassed(caseData.interestedPartyRepsDueDate) &&
+		deadlineHasPassed(caseData.finalCommentsDueDate) &&
+		!deadlineHasPassed(caseData.proofsOfEvidenceDueDate) &&
+		!caseData.LPAProofsSubmittedDate
 	);
 };
 
@@ -411,6 +474,22 @@ const isAppellantFinalCommentDue = (appealCaseData) => {
 };
 
 /**
+ * @param {AppealCaseDetailed} caseData
+ * @param {UserRole} userType
+ * @returns {boolean}
+ */
+const shouldDisplayFinalCommentsDueBannerAppellant = (caseData, userType) => {
+	return (
+		!!caseData.finalCommentsDueDate &&
+		userType === APPEAL_USER_ROLES.APPELLANT &&
+		deadlineHasPassed(caseData.statementDueDate) &&
+		deadlineHasPassed(caseData.interestedPartyRepsDueDate) &&
+		!deadlineHasPassed(caseData.finalCommentsDueDate) &&
+		!caseData.appellantCommentsSubmittedDate
+	);
+};
+
+/**
  * @param {AppealCaseDetailed} appealCaseData return object from database call
  * @returns {boolean}
  */
@@ -419,6 +498,23 @@ const isAppellantProofsOfEvidenceDue = (appealCaseData) => {
 		!!appealCaseData.proofsOfEvidenceDueDate &&
 		!appealCaseData.appellantProofsSubmittedDate &&
 		appealCaseData.caseStatus === APPEAL_CASE_STATUS.EVIDENCE
+	);
+};
+
+/**
+ * @param {AppealCaseDetailed} caseData
+ * @param {UserRole} userType
+ * @returns {boolean}
+ */
+const shouldDisplayProofEvidenceDueBannerAppellant = (caseData, userType) => {
+	return (
+		!!caseData.proofsOfEvidenceDueDate &&
+		userType === APPEAL_USER_ROLES.APPELLANT &&
+		deadlineHasPassed(caseData.statementDueDate) &&
+		deadlineHasPassed(caseData.interestedPartyRepsDueDate) &&
+		deadlineHasPassed(caseData.finalCommentsDueDate) &&
+		!deadlineHasPassed(caseData.proofsOfEvidenceDueDate) &&
+		!caseData.appellantProofsSubmittedDate
 	);
 };
 
@@ -481,6 +577,21 @@ const isRule6StatementDue = (appealCaseData) => {
 };
 
 /**
+ * @param {AppealCaseDetailed} caseData
+ * @param {UserRole} userType
+ * @returns {boolean}
+ */
+const shouldDisplayStatementsDueBannerRule6 = (caseData, userType) => {
+	return (
+		!!caseData.statementDueDate &&
+		userType === APPEAL_USER_ROLES.RULE_6_PARTY &&
+		!!caseData.lpaQuestionnairePublishedDate &&
+		!representationExists(caseData.Representations, REPRESENTATION_TYPES.STATEMENT, true) &&
+		!deadlineHasPassed(caseData.statementDueDate)
+	);
+};
+
+/**
  * @param {AppealCaseDetailed} appealCaseData return object from database call
  * @returns {boolean}
  */
@@ -493,6 +604,23 @@ const isRule6ProofOfEvidenceDue = (appealCaseData) => {
 			true
 		) &&
 		appealCaseData.caseStatus === APPEAL_CASE_STATUS.EVIDENCE
+	);
+};
+
+/**
+ * @param {AppealCaseDetailed} caseData
+ * @param {UserRole} userType
+ * @returns {boolean}
+ */
+const shouldDisplayProofEvidenceDueBannerRule6 = (caseData, userType) => {
+	return (
+		!!caseData.proofsOfEvidenceDueDate &&
+		userType === APPEAL_USER_ROLES.RULE_6_PARTY &&
+		deadlineHasPassed(caseData.statementDueDate) &&
+		deadlineHasPassed(caseData.interestedPartyRepsDueDate) &&
+		deadlineHasPassed(caseData.finalCommentsDueDate) &&
+		!deadlineHasPassed(caseData.proofsOfEvidenceDueDate) &&
+		!representationExists(caseData.Representations, REPRESENTATION_TYPES.PROOFS_OF_EVIDENCE, true)
 	);
 };
 
@@ -538,5 +666,13 @@ module.exports = {
 	mapToRule6DashboardDisplayData,
 	isToDoRule6Dashboard,
 	representationExists,
-	representationPublished
+	representationPublished,
+	shouldDisplayQuestionnaireDueNotification,
+	shouldDisplayStatementsDueBannerLPA,
+	shouldDisplayStatementsDueBannerRule6,
+	shouldDisplayFinalCommentsDueBannerLPA,
+	shouldDisplayFinalCommentsDueBannerAppellant,
+	shouldDisplayProofEvidenceDueBannerAppellant,
+	shouldDisplayProofEvidenceDueBannerLPA,
+	shouldDisplayProofEvidenceDueBannerRule6
 };
